Add simple polyphonic synth
This commit adds the code for a functional polyphonic synth with
a configurable number of voices from 1 to 16.

The synth receives midi imput via USB. Only noteon and noteoff
messages are handled. Velocity is used to determine wave
amplitude.

This code is a good boilerplate for any teensy polyphonic
synth.